api/inference/requirements.txt
Flask
gunicorn
joblib
tensorflow==2.15.0
opencv-python==4.9.0.80
numpy==1.26.4
matplotlib==3.8.3
pillow==10.2.0